+++ src/app/(payload)/admin/importMap.js
@@ -1,14 +1,24 @@
+import { PayloadListInfo as PayloadListInfo_7ff3c83942f86e4d0dadc6be04c1c4b5 } from '@/components/payload/payload-admin-shell'
+import { PayloadListSubheader as PayloadListSubheader_7ff3c83942f86e4d0dadc6be04c1c4b5 } from '@/components/payload/payload-admin-shell'
 import { PayloadAdminLogo as PayloadAdminLogo_79108ba5474e00f2abe107cc5c104d5d } from '@/components/payload/payload-auth-brand'
+import { PayloadDashboardInfo as PayloadDashboardInfo_7ff3c83942f86e4d0dadc6be04c1c4b5 } from '@/components/payload/payload-admin-shell'
 import { PayloadDashboardLink as PayloadDashboardLink_79108ba5474e00f2abe107cc5c104d5d } from '@/components/payload/payload-auth-brand'
+import { PayloadDashboardHeader as PayloadDashboardHeader_7ff3c83942f86e4d0dadc6be04c1c4b5 } from '@/components/payload/payload-admin-shell'
 import { PayloadLoginIntro as PayloadLoginIntro_79108ba5474e00f2abe107cc5c104d5d } from '@/components/payload/payload-auth-brand'
+import { PayloadNavHeader as PayloadNavHeader_7ff3c83942f86e4d0dadc6be04c1c4b5 } from '@/components/payload/payload-admin-shell'
 import { S3ClientUploadHandler as S3ClientUploadHandler_f97aa6c64367fa259c5bc0567239ef24 } from '@payloadcms/storage-s3/client'
 import { CollectionCards as CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1 } from '@payloadcms/next/rsc'
 
 /** @type import('payload').ImportMap */
 export const importMap = {
+  "@/components/payload/payload-admin-shell#PayloadListInfo": PayloadListInfo_7ff3c83942f86e4d0dadc6be04c1c4b5,
+  "@/components/payload/payload-admin-shell#PayloadListSubheader": PayloadListSubheader_7ff3c83942f86e4d0dadc6be04c1c4b5,
   "@/components/payload/payload-auth-brand#PayloadAdminLogo": PayloadAdminLogo_79108ba5474e00f2abe107cc5c104d5d,
+  "@/components/payload/payload-admin-shell#PayloadDashboardInfo": PayloadDashboardInfo_7ff3c83942f86e4d0dadc6be04c1c4b5,
   "@/components/payload/payload-auth-brand#PayloadDashboardLink": PayloadDashboardLink_79108ba5474e00f2abe107cc5c104d5d,
+  "@/components/payload/payload-admin-shell#PayloadDashboardHeader": PayloadDashboardHeader_7ff3c83942f86e4d0dadc6be04c1c4b5,
   "@/components/payload/payload-auth-brand#PayloadLoginIntro": PayloadLoginIntro_79108ba5474e00f2abe107cc5c104d5d,
+  "@/components/payload/payload-admin-shell#PayloadNavHeader": PayloadNavHeader_7ff3c83942f86e4d0dadc6be04c1c4b5,
   "@payloadcms/storage-s3/client#S3ClientUploadHandler": S3ClientUploadHandler_f97aa6c64367fa259c5bc0567239ef24,
   "@payloadcms/next/rsc#CollectionCards": CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1
 }